requirements.txt
praw
tweepy
gspread
oauth2client
speechrecognition
pyaudio
python-dotenv
